Rename drift.ts to detectSpecPcShapeDrift.ts
